# CUDA: faster IQ2_K, IQ2_KS, IQ2_K_R4