--- a/ggml/src/ggml-cuda/mmq.cuh
+++ b/ggml/src/ggml-cuda/mmq.cuh
@@ -2554,15 +2554,6 @@ template <int mmq_y, int nwarps, bool need_check> static __device__ __forceinlin
     }
 }
 
-#ifdef __CUDA_ARCH__
-static __device__ __forceinline__ int2 get_int_from_table_8(const int & q4, const int8_t * values) {
-    const uint32_t * values32 = (const uint32_t *)values;
-    uint32_t v1 = __byte_perm(values32[0], values32[1], q4);
-    uint32_t v2 = __byte_perm(values32[0], values32[1], q4 >> 16);
-    return make_int2(__byte_perm(v1, v2, 0x6420), __byte_perm(v1, v2, 0x7531));
-}
-#endif
-
 template <int mmq_y, int nwarps, bool need_check> static __device__ __forceinline__ void load_tiles_iq2_ks(
     const char * __restrict__ x, int * __restrict__ x_tile, const int & kbx0, const int & i_max, const int & stride) {